% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oster_functions.R
\name{bound}
\alias{bound}
\title{Delta bound}
\usage{
bound(bo_m_bt, rt_m_ro_t_syy, rm_m_rt_t_syy, beta_tilde, beta, t_x, sigma_xx)
}
\arguments{
\item{...}{arguments passed via \code{oster()}}
}
\value{
numeric for delta
}
\description{
Solves for delta. Used in \code{oster()}
}
\examples{
...
}
