% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oster_functions.R
\name{d1quadsol}
\alias{d1quadsol}
\title{Solve for beta with delta eq to 1}
\usage{
d1quadsol(
  rm_m_rt_t_syy,
  rt_m_ro_t_syy,
  bo_m_bt,
  sigma_xx,
  t_x,
  beta_o,
  beta_tilde
)
}
\arguments{
\item{...}{arguments passed via \code{oster()}}
}
\value{
list containing all solutions, their distances from beta_tilde and
indicator of whether the solution goes in opposite direction of
observed bias
}
\description{
Solves for beta. Used in \code{oster()}
}
\examples{
...
}
