% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oster_functions.R
\name{dnot1cubsol}
\alias{dnot1cubsol}
\title{Solve for beta with delta neq to 1}
\usage{
dnot1cubsol(
  bo_m_bt,
  sigma_xx,
  delta,
  t_x,
  rm_m_rt_t_syy,
  rt_m_ro_t_syy,
  beta_tilde,
  beta_o
)
}
\arguments{
\item{...}{arguments passed via \code{oster()}}
}
\value{
list containing all solutions, their distances from beta_tilde and
indicator of whether the solution goes in opposite direction of
observed bias
}
\description{
Solves for beta. Used in \code{oster()}
}
\examples{
...
}
