% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nunn}
\alias{nunn}
\title{The Slave Trade and the Origins of Mistrust in Africa}
\format{
A data frame with 21822 rows and 60 variables:
\describe{
\item{\code{trust_relatives}}{numeric 0-3}
\item{\code{more variables}}{...}
...
}
}
\source{
\url{https://www.aeaweb.org/articles?id=10.1257/aer.101.7.3221}
}
\usage{
nunn
}
\description{
A dataset containing data from Nunn & Wantchekon (2011) used by Oster (2019)
}
\references{
Nunn, N., & Wantchekon, L. (2011). The slave trade and the
origins of mistrust in Africa. American Economic Review, 101(7), 3221-52.

Oster, E. (2019). Unobservable selection and coefficient
stability: Theory and evidence. Journal of Business & Economic
Statistics, 37(2), 187-204.
}
\keyword{datasets}
