% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oster_functions.R
\name{rbound}
\alias{rbound}
\title{R-squared bound}
\usage{
rbound(
  bo_m_bt,
  rt_m_ro_t_syy,
  beta_tilde,
  beta,
  t_x,
  sigma_xx,
  sigma_yy,
  delta,
  r_tilde
)
}
\arguments{
\item{...}{arguments passed via \code{oster()}}
}
\value{
numeric for maximum r-squared
}
\description{
Solves for maximum r-squared. Used in \code{oster()}
}
\examples{
...
}
