% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oster_functions.R
\name{weighted.var}
\alias{weighted.var}
\title{Compute Weighted Variance}
\usage{
weighted.var(x, w, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector of values}

\item{w}{a numeric vector of weights}
}
\value{
numeric for weighted variance
}
\description{
what it says
}
\examples{
...
}
