% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oster_functions.R
\name{within.rsq}
\alias{within.rsq}
\title{Compute within R-squared}
\usage{
\method{within}{rsq}(lm.fit, fe)
}
\arguments{
\item{lm.fit}{an object of class \code{lm}.}

\item{fe}{character vector specifying the variable names for the fixed effects}
}
\value{
numeric for within r-squared.
}
\description{
what it says
}
\examples{
...
}
